requirements.txt
asgiref==3.8.1
Django==5.1.1
ruff==0.6.8
sqlparse==0.5.1

python-socketio~=5.11.4